add authorization for account

+++ db/query/account.sql
@@ -16,23 +16,24 @@ FOR NO KEY UPDATE;
 
 -- name: ListAccount :many
 SELECT * FROM accounts
+WHERE owner = $1
 ORDER BY id
-LIMIT $1
-OFFSET $2;
+LIMIT $2
+OFFSET $3;
 
 
 -- name: UpdateAccount :one
 UPDATE accounts
-set balance= $2
-WHERE id = $1
+set balance= $3
+WHERE id = $1 AND owner = $2
 RETURNING *;
 
 -- name: AddAccountBalance :one
 UPDATE accounts
 set balance= balance + sqlc.arg(amount)
-WHERE id = sqlc.arg(id)
+WHERE id = sqlc.arg(id) 
 RETURNING *;
 
 
 -- name: DeleteAccount :exec
-DELETE FROM accounts WHERE id = $1;
+DELETE FROM accounts WHERE id = $1 AND owner = $2;
